Optimization of DBA
I noticed that the actual DBA function was calling the medoid_index of sequences, which was calling the sum_of_squares(sequences{i}, seuences), which in the end was calling the dtw (sequences{i], sequences{j}). So, basically the algorithm was doing dtw(S,T) and then dtw(T,S) [and dtw(S,T) = dtw(T,S)].
To optimize the DBA code, the previous medoid-index and SumOfSquares function were written as a script, allowing to store the value dtw value that was computed for each “pair of sequences{i}”. 

Two for loops that go through all the set of pulses were created, where i is the index of the external and j of the internal loop, when i=j nothing happens because the dtw would be 0 so it isn't necessary to compute it. Now, when i < j the dtw(sequences{i},sequences{j}) value wasn't calculated yet, so it is computed and stored in a matrix. But when i > j the dtw(sequences{j},sequences{i}) already exists so the algorithm goes to the matrix and just picks the value in the position (j,i).  And in the end this reduced a lot the computing time.

Besides that, I noticed that the dtw function implemented in the algorithm was slower than the Matlab intrinsic one.